# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Bot Hosting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
  /platform/config:
    get:
      operationId: getPlatformConfig
      tags: [platform]
      summary: Get current platform configuration (configured bot + slot info)
      responses:
        "200":
          description: Platform config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformConfig"
        "503":
          description: Platform not initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePlatformConfig
      tags: [platform]
      summary: Update platform's configured bot and/or slot count (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlatformConfigBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid bot repo or shrinking below occupied slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin auth required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /platform/login:
    post:
      operationId: platformLogin
      tags: [platform]
      summary: Verify admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              required: [password]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /github/fetch-app-json:
    post:
      operationId: fetchAppJson
      tags: [github]
      summary: Fetch and parse app.json from a GitHub repo URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchAppJsonBody"
      responses:
        "200":
          description: Parsed app.json data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppJsonData"
        "400":
          description: Invalid URL or missing app.json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    FetchAppJsonBody:
      type: object
      properties:
        repoUrl:
          type: string
          description: GitHub repository URL (https://github.com/owner/repo)
      required: [repoUrl]
    AppJsonEnvVar:
      type: object
      properties:
        description:
          type: string
        value:
          type: string
        required:
          type: boolean
        generator:
          type: string
      required: []
    AppJsonData:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        keywords:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AppJsonEnvVar"
        scripts:
          type: object
          additionalProperties:
            type: string
      required: [name]
...
    PlatformConfig:
      type: object
      properties:
        botRepoUrl:
          type: string
        botRepoOwner:
          type: string
        botRepoName:
          type: string
        botName:
          type: string
        botDescription:
          type: string
          nullable: true
        botLogo:
          type: string
          nullable: true
        botAppJson:
          $ref: "#/components/schemas/AppJsonData"
        slotCount:
          type: integer
        totalSlots:
          type: integer
        occupiedSlots:
          type: integer
        availableSlots:
          type: integer
        updatedAt:
          type: string
          format: date-time
        adminPasswordIsDefault:
          type: boolean
      required:
        [botRepoUrl, botRepoOwner, botRepoName, botName, botAppJson, slotCount, totalSlots, occupiedSlots, availableSlots, updatedAt, adminPasswordIsDefault]
    UpdatePlatformConfigBody:
      type: object
      properties:
        botRepoUrl:
          type: string
          description: New GitHub repo URL for the bot
        slotCount:
          type: integer
          minimum: 1
          maximum: 500
...